+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming
 
 - test: make sure pytest exits completely after running async tests
+- refactor: in `_wait_for_store_to_finish`, instead of waiting with `asyncio.sleep`, run the store event loop when conditions are not satisfied
 
 ## Version 0.21.1